requirements.txt
Flask==2.2.2
opencv-python==4.9.0.80
Pillow==9.4.0
numpy==1.24.3
img2vec-pytorch==1.0.1
tensorflow==2.16.1
scikit-learn==1.3.0
requests-mock
fuzzytm
blosc2
cython
clyent==1.2.1
nbformat==5.4.0
requests==2.28.1
black>=24.4.2
fsspec==2023.3.0
setuptools >=58.0.0